+++ src/pages/steps/AdditionalInformationStep/OrganizerPicker.tsx
@@ -8,8 +8,8 @@ import { useGetOffersByCreatorQuery } from '@/hooks/api/offers';
 import { useGetOrganizersByQueryQuery } from '@/hooks/api/organizers';
 import { useGetUserQuery } from '@/hooks/api/user';
 import { SupportedLanguages } from '@/i18n/index';
-import { CultuurKuurIcon } from '@/pages/CultuurKuurIcon';
-import { UitpasIcon } from '@/pages/UitpasIcon';
+import { CultuurKuurIcon } from '@/ui/CultuurKuurIcon';
+import { UitpasIcon } from '@/ui/UitpasIcon';
 import { Organizer } from '@/types/Organizer';
 import { Values } from '@/types/Values';
 import { Alert, AlertVariants } from '@/ui/Alert';
